Add new documentation manager script
This commit adds the script at the `tools/documentation-manager.js`,
which is needed in order to manipulate the project's documentation.
It can build and clean the documentation for the source code.
The documentation manager generates the documentation in the Markdown
format. Output goes to the `docs/documentation` directory.

This commit also changes the npm script in the `package.json`
configuration file.